scripts/requirements_bench.txt
openai-whisper
datasets
jiwer
soundfile
librosa
torch
tqdm
